Resolve maxDirtyPagesRatio once; moved DFLT_MAX_DIRTY_PAGES_RATIO to PageMemoryImpl;
