+++ modules/core/src/test/java/org/apache/ignite/internal/processors/cache/persistence/pagemem/BPlusTreePageMemoryImplTest.java
@@ -119,7 +119,8 @@ public class BPlusTreePageMemoryImplTest extends BPlusTreeSelfTest {
             () -> true,
             new DataRegionMetricsImpl(new DataRegionConfiguration(), cctx),
             PageMemoryImpl.ThrottlingPolicy.DISABLED,
-            clo
+            clo,
+            PageMemoryImpl.DFLT_MAX_DIRTY_PAGES_RATIO
         );
 
         mem.start();
