+++ modules/core/src/test/java/org/apache/ignite/internal/processors/cache/persistence/pagemem/IndexStoragePageMemoryImplTest.java
@@ -130,7 +130,8 @@ public class IndexStoragePageMemoryImplTest extends IndexStorageSelfTest {
             () -> true,
             new DataRegionMetricsImpl(new DataRegionConfiguration(), cctx),
             PageMemoryImpl.ThrottlingPolicy.DISABLED,
-            clo
+            clo,
+            PageMemoryImpl.DFLT_MAX_DIRTY_PAGES_RATIO
         );
     }
 }
