+++ modules/core/src/test/java/org/apache/ignite/internal/processors/cache/persistence/pagemem/PageMemoryImplNoLoadTest.java
@@ -120,7 +120,8 @@ public class PageMemoryImplNoLoadTest extends PageMemoryNoLoadSelfTest {
             () -> true,
             new DataRegionMetricsImpl(new DataRegionConfiguration(), cctx),
             PageMemoryImpl.ThrottlingPolicy.DISABLED,
-            clo
+            clo,
+            PageMemoryImpl.DFLT_MAX_DIRTY_PAGES_RATIO
         );
     }
